refactor tests to reduce redundancy
