+++ queryall-api/src/main/java/org/queryall/api/namespace/NamespaceEntrySchema.java
@@ -253,8 +253,9 @@ public class NamespaceEntrySchema extends QueryAllSchema
             con.setAutoCommit(false);
             
             con.add(NamespaceEntrySchema.getNamespaceTypeUri(), RDF.TYPE, OWL.CLASS, contextKeyUri);
-            con.add(NamespaceEntrySchema.getNamespaceTypeUri(), RDFS.SUBCLASSOF, f.createURI("http://purl.org/court/def/2009/coin#", "URISpace"), contextKeyUri);
-                     
+            con.add(NamespaceEntrySchema.getNamespaceTypeUri(), RDFS.SUBCLASSOF,
+                    f.createURI("http://purl.org/court/def/2009/coin#", "URISpace"), contextKeyUri);
+            
             con.add(NamespaceEntrySchema.getNamespacePreferredPrefix(), RDF.TYPE, OWL.DATATYPEPROPERTY, contextKeyUri);
             con.add(NamespaceEntrySchema.getNamespacePreferredPrefix(), RDFS.SUBPROPERTYOF, Constants.DC_TITLE,
                     contextKeyUri);
@@ -298,7 +299,8 @@ public class NamespaceEntrySchema extends QueryAllSchema
                     contextKeyUri);
             
             con.add(NamespaceEntrySchema.getNamespaceUriTemplate(), RDF.TYPE, OWL.DATATYPEPROPERTY, contextKeyUri);
-            con.add(NamespaceEntrySchema.getNamespaceUriTemplate(), RDFS.SUBPROPERTYOF, f.createURI("http://purl.org/court/def/2009/coin#", "uriTemplate"), contextKeyUri);
+            con.add(NamespaceEntrySchema.getNamespaceUriTemplate(), RDFS.SUBPROPERTYOF,
+                    f.createURI("http://purl.org/court/def/2009/coin#", "uriTemplate"), contextKeyUri);
             con.add(NamespaceEntrySchema.getNamespaceUriTemplate(), RDFS.RANGE, RDFS.LITERAL, contextKeyUri);
             con.add(NamespaceEntrySchema.getNamespaceUriTemplate(), RDFS.DOMAIN,
                     NamespaceEntrySchema.getNamespaceTypeUri(), contextKeyUri);
@@ -308,7 +310,8 @@ public class NamespaceEntrySchema extends QueryAllSchema
                     contextKeyUri);
             
             con.add(NamespaceEntrySchema.getNamespaceSeparator(), RDF.TYPE, OWL.DATATYPEPROPERTY, contextKeyUri);
-            con.add(NamespaceEntrySchema.getNamespaceSeparator(), RDFS.SUBPROPERTYOF, f.createURI("http://purl.org/court/def/2009/coin#", "fragmentSeparator"), contextKeyUri);
+            con.add(NamespaceEntrySchema.getNamespaceSeparator(), RDFS.SUBPROPERTYOF,
+                    f.createURI("http://purl.org/court/def/2009/coin#", "fragmentSeparator"), contextKeyUri);
             con.add(NamespaceEntrySchema.getNamespaceSeparator(), RDFS.RANGE, RDFS.LITERAL, contextKeyUri);
             con.add(NamespaceEntrySchema.getNamespaceSeparator(), RDFS.DOMAIN,
                     NamespaceEntrySchema.getNamespaceTypeUri(), contextKeyUri);
@@ -316,8 +319,6 @@ public class NamespaceEntrySchema extends QueryAllSchema
                     RDFS.LABEL,
                     f.createLiteral("This value is used as the separator between the namespace and the identifier portion of the URI."),
                     contextKeyUri);
-            
-            
             
             if(modelVersion == 1)
             {
