+++ queryall-api/src/test/java/org/queryall/api/test/AbstractNormalisationRuleTest.java
@@ -25,7 +25,7 @@ import org.queryall.exception.InvalidStageException;
  * 
  * @author Peter Ansell p_ansell@yahoo.com
  */
-public abstract class AbstractNormalisationRuleTest
+public abstract class AbstractNormalisationRuleTest extends AbstractProfilableNormalisationRuleTest
 {
     private URI testStageInvalidInclusionRuleUri;
     private URI testStageAllValidAndInvalidRuleUri;
@@ -38,11 +38,18 @@ public abstract class AbstractNormalisationRuleTest
      */
     public abstract Set<URI> getExpectedValidStages();
     
+    @Override
+    public final NormalisationRule getNewTestProfilable()
+    {
+        return this.getNewTestRule();
+    }
+    
     /**
      * Create a new profile instance with default properties
      * 
      * @return A new profile instance with default properties
      */
+    @Override
     public abstract Profile getNewTestProfile();
     
     /**
@@ -55,9 +62,12 @@ public abstract class AbstractNormalisationRuleTest
     /**
      * @throws java.lang.Exception
      */
+    @Override
     @Before
     public void setUp() throws Exception
     {
+        super.setUp();
+        
         final ValueFactory f = new MemValueFactory();
         
         // this.testTrueSparqlNormalisationRuleUri =
@@ -90,9 +100,12 @@ public abstract class AbstractNormalisationRuleTest
     /**
      * @throws java.lang.Exception
      */
+    @Override
     @After
     public void tearDown() throws Exception
     {
+        super.tearDown();
+        
         // this.testTrueSparqlNormalisationRuleUri = null;
         // this.testFalseSparqlNormalisationRuleUri = null;
         this.testStageInvalidInclusionRuleUri = null;
@@ -111,12 +124,16 @@ public abstract class AbstractNormalisationRuleTest
         
         final Collection<URI> includedStages = this.validStages;
         
-        Assert.assertNotNull("Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places", this.validStages);
-
+        Assert.assertNotNull(
+                "Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places",
+                this.validStages);
+        
         final Collection<URI> excludedStages = this.invalidStages;
         
-        Assert.assertNotNull("Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places", this.invalidStages);
-
+        Assert.assertNotNull(
+                "Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places",
+                this.invalidStages);
+        
         for(final URI nextIncludedStage : includedStages)
         {
             try
@@ -150,15 +167,20 @@ public abstract class AbstractNormalisationRuleTest
         
         final Collection<URI> excludedStages = this.invalidStages;
         
-        Assert.assertNotNull("Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places", this.validStages);
-        Assert.assertNotNull("Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places", this.invalidStages);
-
+        Assert.assertNotNull(
+                "Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places",
+                this.validStages);
+        Assert.assertNotNull(
+                "Unexpected null stages. Check if super.setUp() and super.tearDown() are called in the relevant places",
+                this.invalidStages);
+        
         for(final URI nextInvalidStage : this.invalidStages)
         {
             try
             {
                 normalisationRule.addStage(nextInvalidStage);
-                Assert.fail("Did not find expected invalid stage exception for setStages nextInvalidStage="+nextInvalidStage);
+                Assert.fail("Did not find expected invalid stage exception for setStages nextInvalidStage="
+                        + nextInvalidStage);
             }
             catch(final InvalidStageException ise)
             {
