--- queryall-api/src/test/java/org/queryall/api/test/AbstractSpinInferencingRuleTest.java
@@ -1,148 +1,0 @@
-/**
- * 
- */
-package org.queryall.api.test;
-
-import java.util.HashSet;
-import java.util.Set;
-
-import org.junit.After;
-import org.junit.Before;
-import org.openrdf.model.URI;
-import org.openrdf.model.ValueFactory;
-import org.openrdf.model.impl.ValueFactoryImpl;
-import org.openrdf.repository.Repository;
-import org.openrdf.repository.RepositoryConnection;
-import org.openrdf.repository.RepositoryException;
-import org.openrdf.repository.sail.SailRepository;
-import org.openrdf.sail.memory.MemoryStore;
-import org.queryall.api.rdfrule.NormalisationRuleSchema;
-import org.queryall.api.rdfrule.SpinInferencingRule;
-import org.queryall.api.rdfrule.SpinNormalisationRule;
-
-/**
- * Abstract unit test for SpinInferencingRule API.
- * 
- * @author Peter Ansell p_ansell@yahoo.com
- */
-public abstract class AbstractSpinInferencingRuleTest extends AbstractSpinNormalisationRuleTest
-{
-    /**
-     * 
-     * @param sparqlPrefixes
-     * @param constructGraphPattern
-     * @param wherePattern
-     * @return
-     */
-    private static String mergeQuery(final String sparqlPrefixes, final String wherePattern)
-    {
-        return new StringBuilder(sparqlPrefixes).append(" ASK WHERE { ").append(wherePattern).append(" }").toString();
-    }
-    
-    private URI testMultipleWherePatternsSparqlNormalisationRuleUri;
-    private URI testStageEmptyConstructQuerySetSparqlNormalisationRuleUri;
-    private String testStartingUriAEOBase;
-    private String testFinalUriAEOBase;
-    
-    private String testStartingUriPOBase;
-    
-    private String testFinalUriPOBase;
-    private Repository testRepository;
-    private RepositoryConnection testRepositoryConnection;
-    private ValueFactory testValueFactory;
-    
-    @Override
-    public final Set<URI> getExpectedValidStages()
-    {
-        final Set<URI> results = new HashSet<URI>();
-        
-        results.add(NormalisationRuleSchema.getRdfruleStageAfterQueryParsing());
-        results.add(NormalisationRuleSchema.getRdfruleStageAfterResultsImport());
-        results.add(NormalisationRuleSchema.getRdfruleStageAfterResultsToPool());
-        
-        return results;
-    }
-    
-    /**
-     * Create a new instance of the SpinInferencingRule implementation being tested.
-     * 
-     * @return a new instance of the implemented SpinInferencingRule
-     */
-    public abstract SpinInferencingRule getNewTestSpinInferencingRule();
-    
-    /**
-     * Final method, so that implementing test cases must supply a SpinInferencingRule instead,
-     * through getNewTestSpinInferencingRule.
-     * 
-     * @return A NormalisationRule that is also a SpinNormalisationRule
-     */
-    @Override
-    public final SpinNormalisationRule getNewTestSpinNormalisationRule()
-    {
-        return this.getNewTestSpinInferencingRule();
-    }
-    
-    /**
-     * @throws java.lang.Exception
-     */
-    @Override
-    @Before
-    public void setUp() throws Exception
-    {
-        super.setUp();
-        
-        this.testRepository = new SailRepository(new MemoryStore());
-        this.testRepository.initialize();
-        this.testRepositoryConnection = this.testRepository.getConnection();
-        this.testValueFactory = new ValueFactoryImpl();
-        
-        this.testMultipleWherePatternsSparqlNormalisationRuleUri =
-                this.testValueFactory.createURI("http://example.org/test/multipleWherePatternsSparqlNormalisationRule");
-        this.testStageEmptyConstructQuerySetSparqlNormalisationRuleUri =
-                this.testValueFactory
-                        .createURI("http://example.org/test/emptyConstructQuerySetSparqlNormalisationRule");
-        
-        this.testStartingUriAEOBase = "http://purl.obolibrary.org/obo/AEO_";
-        this.testFinalUriAEOBase = "http://bio2rdf.org/obo_aeo:";
-        this.testStartingUriPOBase = "http://purl.obolibrary.org/obo/PO_";
-        this.testFinalUriPOBase = "http://bio2rdf.org/obo_po:";
-        
-    }
-    
-    /**
-     * @throws java.lang.Exception
-     */
-    @Override
-    @After
-    public void tearDown() throws Exception
-    {
-        super.tearDown();
-        
-        if(this.testRepositoryConnection != null)
-        {
-            try
-            {
-                this.testRepositoryConnection.close();
-            }
-            catch(final RepositoryException ex)
-            {
-                ex.printStackTrace();
-            }
-            finally
-            {
-                this.testRepositoryConnection = null;
-            }
-        }
-        
-        this.testRepository = null;
-        this.testValueFactory = null;
-        
-        this.testMultipleWherePatternsSparqlNormalisationRuleUri = null;
-        
-        this.testStartingUriAEOBase = null;
-        this.testFinalUriAEOBase = null;
-        this.testStartingUriPOBase = null;
-        this.testFinalUriPOBase = null;
-    }
-    
-}
