--- queryall-ext-provider-http/src/test/java/org/queryall/impl/provider/test/ProfilableHttpOnlyProviderImplTest.java
@@ -1,31 +1,0 @@
-/**
- * 
- */
-package org.queryall.impl.provider.test;
-
-import org.queryall.api.base.ProfilableInterface;
-import org.queryall.api.profile.Profile;
-import org.queryall.api.test.AbstractProfilableProviderTest;
-import org.queryall.impl.profile.ProfileImpl;
-import org.queryall.impl.provider.HttpOnlyProviderImpl;
-
-/**
- * Provides the implementation of the Provider class for the Abstract test class provided with
- * queryall-api-tests.jar
- * 
- * @author Peter Ansell p_ansell@yahoo.com
- */
-public class ProfilableHttpOnlyProviderImplTest extends AbstractProfilableProviderTest
-{
-    @Override
-    public ProfilableInterface getNewTestProfilable()
-    {
-        return new HttpOnlyProviderImpl();
-    }
-    
-    @Override
-    public Profile getNewTestProfile()
-    {
-        return new ProfileImpl();
-    }
-}
