--- queryall-ext-querytype-rdf/src/test/java/org/queryall/impl/querytype/test/ProfilableRdfInputQueryTypeImplTest.java
@@ -1,31 +1,0 @@
-/**
- * 
- */
-package org.queryall.impl.querytype.test;
-
-import org.queryall.api.base.ProfilableInterface;
-import org.queryall.api.profile.Profile;
-import org.queryall.api.test.AbstractProfilableQueryTypeTest;
-import org.queryall.impl.profile.ProfileImpl;
-import org.queryall.impl.querytype.RdfInputQueryTypeImpl;
-
-/**
- * Provides the implementation of the Provider class for the Abstract test class provided with
- * queryall-api-tests.jar
- * 
- * @author Peter Ansell p_ansell@yahoo.com
- */
-public class ProfilableRdfInputQueryTypeImplTest extends AbstractProfilableQueryTypeTest
-{
-    @Override
-    public ProfilableInterface getNewTestProfilable()
-    {
-        return new RdfInputQueryTypeImpl();
-    }
-    
-    @Override
-    public Profile getNewTestProfile()
-    {
-        return new ProfileImpl();
-    }
-}
