--- queryall-ext-querytype-rdf/src/test/java/org/queryall/impl/querytype/test/RdfInputQueryTypeImpl3Test.java
@@ -1,21 +1,0 @@
-/**
- * 
- */
-package org.queryall.impl.querytype.test;
-
-import org.queryall.api.querytype.RdfInputQueryType;
-import org.queryall.api.test.AbstractRdfInputQueryTypeTest;
-import org.queryall.impl.querytype.RdfInputQueryTypeImpl;
-
-/**
- * @author Peter Ansell p_ansell@yahoo.com
- */
-public class RdfInputQueryTypeImpl3Test extends AbstractRdfInputQueryTypeTest
-{
-    @Override
-    public RdfInputQueryType getNewTestRdfInputQueryType()
-    {
-        return new RdfInputQueryTypeImpl();
-    }
-    
-}
