+++ queryall-ext-querytype-regex/src/test/java/org/queryall/impl/querytype/test/RegexInputQueryTypeImpl3Test.java
@@ -3,8 +3,10 @@
  */
 package org.queryall.impl.querytype.test;
 
+import org.queryall.api.profile.Profile;
 import org.queryall.api.querytype.RegexInputQueryType;
 import org.queryall.api.test.AbstractRegexInputQueryTypeTest;
+import org.queryall.impl.profile.ProfileImpl;
 import org.queryall.impl.querytype.RegexInputQueryTypeImpl;
 
 /**
@@ -13,9 +15,14 @@ import org.queryall.impl.querytype.RegexInputQueryTypeImpl;
 public class RegexInputQueryTypeImpl3Test extends AbstractRegexInputQueryTypeTest
 {
     @Override
+    public Profile getNewTestProfile()
+    {
+        return new ProfileImpl();
+    }
+    
+    @Override
     public RegexInputQueryType getNewTestRegexInputQueryType()
     {
         return new RegexInputQueryTypeImpl();
     }
-    
 }
