--- queryall-ext-querytype-regex/src/test/java/org/queryall/impl/querytype/test/RegexInputQueryTypeImplTest.java
@@ -1,23 +1,0 @@
-/**
- * 
- */
-package org.queryall.impl.querytype.test;
-
-import org.queryall.api.querytype.QueryType;
-import org.queryall.api.test.AbstractQueryTypeTest;
-import org.queryall.impl.querytype.RegexInputQueryTypeImpl;
-
-/**
- * Provides the implementation of the Provider class for the Abstract test class provided with
- * queryall-api-tests.jar
- * 
- * @author Peter Ansell p_ansell@yahoo.com
- */
-public class RegexInputQueryTypeImplTest extends AbstractQueryTypeTest
-{
-    @Override
-    public QueryType getNewTestQueryType()
-    {
-        return new RegexInputQueryTypeImpl();
-    }
-}
