--- queryall-ext-rule-regex/src/test/java/org/queryall/impl/rule/test/ProfilableRegexNormalisationRuleImplTest.java
@@ -1,30 +1,0 @@
-/**
- * 
- */
-package org.queryall.impl.rule.test;
-
-import org.queryall.api.base.ProfilableInterface;
-import org.queryall.api.profile.Profile;
-import org.queryall.api.test.AbstractProfilableNormalisationRuleTest;
-import org.queryall.impl.profile.ProfileImpl;
-import org.queryall.impl.rdfrule.RegexTransformingRuleImpl;
-
-/**
- * 
- * 
- * @author Peter Ansell p_ansell@yahoo.com
- */
-public class ProfilableRegexNormalisationRuleImplTest extends AbstractProfilableNormalisationRuleTest
-{
-    @Override
-    public ProfilableInterface getNewTestProfilable()
-    {
-        return new RegexTransformingRuleImpl();
-    }
-    
-    @Override
-    public Profile getNewTestProfile()
-    {
-        return new ProfileImpl();
-    }
-}
