+++ queryall-ext-rule-regex/src/test/java/org/queryall/impl/rule/test/RegexTransformingNormalisationRuleImplTest.java
@@ -1,0 +1,31 @@
+/**
+ * 
+ */
+package org.queryall.impl.rule.test;
+
+import org.queryall.api.profile.Profile;
+import org.queryall.api.rdfrule.RegexNormalisationRule;
+import org.queryall.api.test.AbstractRegexNormalisationRuleTest;
+import org.queryall.impl.profile.ProfileImpl;
+import org.queryall.impl.rdfrule.RegexTransformingRuleImpl;
+
+/**
+ * 
+ * 
+ * @author Peter Ansell p_ansell@yahoo.com
+ */
+public class RegexTransformingNormalisationRuleImplTest extends AbstractRegexNormalisationRuleTest
+{
+    
+    @Override
+    public Profile getNewTestProfile()
+    {
+        return new ProfileImpl();
+    }
+    
+    @Override
+    public RegexNormalisationRule getNewTestRegexNormalisationRule()
+    {
+        return new RegexTransformingRuleImpl();
+    }
+}
