+++ queryall-ext-rule-spin/src/test/java/org/queryall/api/test/AbstractSpinConstraintRuleTest.java
